Eclogues.API

Description

Eclogues REST API definition. See also the README.

Documentation

type API = ("jobs" :> Get' [Status]) :<|> (("jobs" :> (Capture "name" Name :> Get' Status)) :<|> (("jobs" :> (Capture "name" Name :> ("stage" :> Get' Stage))) :<|> (("jobs" :> (Capture "name" Name :> ("stage" :> (ReqBody `[JSON]` Stage :> Put `[JSON]` ())))) :<|> (("jobs" :> (Capture "name" Name :> ("scheduler" :> Get' ()))) :<|> (("jobs" :> (Capture "name" Name :> ("output" :> (QueryParam "path" AbsFile :> Get' ())))) :<|> (("jobs" :> (Capture "name" Name :> Delete `[JSON]` ())) :<|> (("jobs" :> (ReqBody `[JSON]` Spec :> Post `[JSON]` ())) :<|> ("health" :> Get' Health))))))))

Eclogues API definition.

data JobError

The various ways requests can fail.

Constructors

JobNameUsed

NoSuchJob
Target job does not exist.

JobMustExist Name
A referenced job does not exist.

JobCannotHaveFailed Name

JobMustBeTerminated Bool

OutstandingDependants [Name]
A job cannot be deleted because some dependants have not terminated.

InvalidStageTransition String
Cannot transition from the current stage to the requested stage.

SchedulerRedirect URI
Not actually an error: used to redirect to a scheduler page.

SchedulerInaccessible
Cannot contact the scheduler.

zkNode :: String

The Zookeeper node on which Eclogues instances run elections.

parseZKData :: ByteString -> Maybe (String, Word16)

Attempt to parse a host and port from a Zookeeper node involved in Eclogues elections.

type AbsFile = Path Abs File

An absolute Path to a file.

type Get' = Get `[JSON]`

Health reporting

data Health

Datatype for health reporting.

mkHealth

Arguments

:: Bool
schedulerAccessible

-> Health

schedulerAccessible :: Lens' Health Bool

Whether Eclogues can access the scheduler.