Eclogues.Job

Description

Specification of jobs and the current state of submitted jobs.

Job status

data Status

Information about a submitted job.

class HasSpec c => HasStatus c where

Minimal complete definition

status

Methods

status :: Lens' c Status

stage :: Lens' c Stage

satis :: Lens' c Satisfiability

uuid :: Lens' c UUID

The UUID used to identify this job in the scheduler.

mkStatus

Arguments

:: Spec
spec

-> Stage
stage

-> Satisfiability
satis

-> UUID
uuid

-> Status

Status not-so-smart constructor.

Job spec

data Spec

Description of a job to run.

class HasResources c => HasSpec c where

Minimal complete definition

spec

Methods

spec :: Lens' c Spec

name :: Lens' c Name

command :: Lens' c Command

The bash command to execute.

outputFiles :: Lens' c [OutputPath]

List of output files where the working directory is the root (/).

captureStdout :: Lens' c Bool

Whether to capture the standard output of the job as a file stdout.

dependsOn :: Lens' c [Name]

The names of the jobs this job requires output from.

mkSpec

Arguments

:: Name
name

-> Command
command

-> Resources
resources

-> [OutputPath]
outputFiles

-> Bool
captureStdout

-> [Name]
dependsOn

-> Spec

Spec not-so-smart constructor.

data Name

A job name.

nameText :: Name -> Text

Name as text.

mkName :: MonadPlus m => Text -> m Name

A Name if the provided text is valid, or otherwise fail.

uuidName :: UUID -> Name

All UUIDs are valid names.

type Command = Text

A bash command.

newtype OutputPath

Newtype wrapper for JSON serialisation purposes.

Constructors

OutputPath

Fields

getOutputPath :: Path Abs File

Resources

data Resources

A set of resources required to run a job.

class HasResources c where

Most of these are Getters to enforce mkResources constraints.

Minimal complete definition

resources

Methods

resources :: Lens' c Resources

disk :: Getter c Data

ram :: Getter c Data

cpu :: Getter c Parallelism

time :: Getter c Time

mkResources :: MonadPlus m => Data -> Data -> Parallelism -> Time -> m Resources

Resources smart constructor. Checks all resource values are positive and under one peta(unit), and fails if not.

Conversions

The following conversion functions are guaranteed to neither overflow nor underflow, due to constraints enforced by mkResources.

diskBytes :: HasResources s => Getter s Int64

diskMB :: HasResources s => Getter s Int64

ramBytes :: HasResources s => Getter s Int64

ramMB :: HasResources s => Getter s Int64

cpuCores :: HasResources s => Getter s Double

timeSeconds :: HasResources s => Getter s Int64

Job satisfiability

data Satisfiability

Whether it's possible to run a job given known constraints.

Constructors

Satisfiable

Unsatisfiable UnsatisfiableReason

SatisfiabilityUnknown
There is insufficient information to determine if the job is satisfiable.

data UnsatisfiableReason

The reason a job was determined to be Unsatisfiable.

Constructors

InsufficientResources
One or more dependencies (dependsOn) are Unsatisfiable.

DependenciesUnsatisfiable [Name]

isUnsatisfiable :: Satisfiability -> Bool

Whether the Satisfiability is Unsatisfiable.

Job lifecycle stage

data Stage

A stage in the lifecycle of a job.

Constructors

Queued QueueStage
Waiting for a node to become available to run.

Waiting Integer
Waiting for n jobs to be Finished.

Running

Killing
In the process of being killed. Transitions to Failed UserKilled.

Finished

Failed FailureReason

RunError RunErrorReason
An internal error occurred.

data QueueStage

Whether the job is on an internal or the scheduler queue.

Constructors

LocalQueue

SchedulerQueue

data FailureReason

The reason a job has Failed.

Constructors

UserKilled
The user requested the job be killed.

NonZeroExitCode Int
The job command exited with a non-zero exit code.

MemoryExceeded
The job exceeded its allocated ram.

DiskExceeded
The job exceeded its allocated disk.

TimeExceeded
The job exceeded its allocated time.

DependencyFailed Name
A job in dependsOn failed.

data RunErrorReason

Details about an internal error.

Constructors

BadSchedulerTransition
The scheduler reported an unexpected state.

SubexecutorFailure
Eclogues subexecutor failure.

SchedulerLost
The scheduler lost the job.

majorStage :: Stage -> String

The Stage constructor name. Could probably be replaced by something from GHC.Generics.

majorStages :: [String]

Names of Stage constructors. Could probably be replaced by something from GHC.Generics.

Predicates

isActiveStage :: Stage -> Bool

isActiveStage = not . 'isTerminationStage'

isTerminationStage :: Stage -> Bool

Whether the stage is terminal. A terminal stage is steady and represents the completion (successful or not) of a job.

isOnScheduler :: Stage -> Bool

Whether the job is known by the remote scheduler.

isPendingStage :: Stage -> Bool

Whether the constructor is queued or waiting on a dependency.

isQueueStage :: Stage -> Bool

Whether the constructor is Queued.